+++ src/wiseagents/cli/wise_agent_cli.py
@@ -3,7 +3,7 @@ import sys
 import threading
 import yaml
 from wiseagents import WiseAgent, WiseAgentMessage, WiseAgentRegistry
-from wiseagents.wise_agent_impl import PassThroughClientAgent, LLMOnlyWiseAgent
+from wiseagents.agents import PassThroughClientAgent, LLMOnlyWiseAgent
 from wiseagents.transports import StompWiseAgentTransport
 
 cond = threading.Condition()
