+++ tests/wiseagents/test_phased_coordinator.py
@@ -2,10 +2,9 @@ import os
 import threading
 
 from wiseagents import WiseAgentMessage, WiseAgentRegistry
-from wiseagents.agents import CollaboratorWiseAgent, PhasedCoordinatorWiseAgent
+from wiseagents.agents import CollaboratorWiseAgent, PassThroughClientAgent, PhasedCoordinatorWiseAgent
 from wiseagents.llm import OpenaiAPIWiseAgentLLM
 from wiseagents.transports import StompWiseAgentTransport
-from wiseagents.wise_agent_impl import PassThroughClientAgent
 
 cond = threading.Condition()
