+++ tests/wiseagents/test_tools_fakeLLM.py
@@ -5,11 +5,11 @@ from typing import Iterable
 
 import pytest
 
+from wiseagents.agents import LLMWiseAgentWithTools, PassThroughClientAgent
 from wiseagents.llm import WiseAgentRemoteLLM
 from wiseagents import WiseAgentMessage, WiseAgentRegistry, WiseAgentTool
 from wiseagents.llm import OpenaiAPIWiseAgentLLM
 from wiseagents.transports.stomp import StompWiseAgentTransport
-from wiseagents.wise_agent_impl import LLMWiseAgentWithTools, PassThroughClientAgent
 
 from openai.types.chat import ChatCompletionMessageParam, ChatCompletion, ChatCompletionToolParam
